refactor: centralize Firebase configuration to prevent multiple initializations; update components to use new Firebase instances and improve error handling with standardized logging
