+++ docs/analise-ia-01/04-plano-acao-implementacao.md
@@ -23,41 +23,40 @@ Plano detalhado para implementação de **33 itens** (15 correções de bugs + 1
 
 ## 🚨 FASE 1: Correções Críticas (Sprints 1-2)
 
-### **Sprint 1: Bugs Críticos**
+### **Sprint 1: Estabilidade e Firebase (8-10h)**
 
-#### **1.1 Inicialização Múltipla do Firebase**
-- **Arquivo**: `src/contexts/AuthContext.jsx`
-- **Tempo estimado**: 4 horas
+#### **1.1 Inicialização Múltipla do Firebase** ⭐ PRIORIDADE MÁXIMA
+- **Arquivo**: `src/contexts/AuthContext.jsx` → `src/firebase/config.js` (novo)
+- **Tempo estimado**: 6 horas (50% buffer incluído)
 - **Responsável**: Dev Senior
-- **Descrição**: Mover inicialização do Firebase para fora do componente
+- **Descrição**: Mover inicialização do Firebase para arquivo separado, evitar múltiplas instâncias
+- **Risco**: Quebra de autenticação durante refator
+- **Mitigação**: Testar em branch separado, backup completo
 - **Implementação**:
   ```jsx
-  // Antes: dentro do componente
-  const app = initializeApp(firebaseConfig)
+  // NOVO: src/firebase/config.js
+  import { initializeApp } from 'firebase/app'
+  import { getAuth } from 'firebase/auth'
+  import { getFirestore } from 'firebase/firestore'
   
-  // Depois: fora do componente
-  const firebaseApp = initializeApp(firebaseConfig)
-  const firebaseAuth = getAuth(firebaseApp)
-  const firebaseDb = getFirestore(firebaseApp)
+  const firebaseConfig = {
+    // ... configuração existente
+  }
+  
+  export const firebaseApp = initializeApp(firebaseConfig)
+  export const firebaseAuth = getAuth(firebaseApp)
+  export const firebaseDb = getFirestore(firebaseApp)
+  
+  // MODIFICADO: src/contexts/AuthContext.jsx
+  import { firebaseAuth, firebaseDb } from '../firebase/config'
   ```
 
-#### **1.2 Dependências Faltantes em useEffect**
-- **Arquivo**: `src/components/RecurrenceConfig.jsx`
-- **Tempo estimado**: 3 horas
-- **Responsável**: Dev Pleno
-- **Descrição**: Corrigir dependências do useEffect para evitar loops
-- **Implementação**:
-  ```jsx
-  useEffect(() => {
-    // ... lógica
-  }, [calculateCurrentRule, onRuleChange]) // Adicionar onRuleChange
-  ```
-
-#### **1.3 Tratamento de Erro Inconsistente**
+#### **1.2 Tratamento de Erro Inconsistente** ⭐ PRIORIDADE ALTA
 - **Arquivo**: `src/hooks/useFirestoreQuery.js`
 - **Tempo estimado**: 2 horas
 - **Responsável**: Dev Pleno
 - **Descrição**: Substituir console.error pelo logger padronizado
+- **Risco**: Baixo - apenas padronização
 - **Implementação**:
   ```jsx
   import { logger } from '../utils/logger'
@@ -65,26 +64,51 @@ Plano detalhado para implementação de **33 itens** (15 correções de bugs + 1
   logger.error('Erro no useFirestoreQuery:', error)
   ```
 
-### **Sprint 2: Bugs de Performance Críticos**
-
-#### **2.1 Re-renders Desnecessários em App.jsx**
-- **Arquivo**: `src/App.jsx`
-- **Tempo estimado**: 6 horas
-- **Responsável**: Dev Senior
-- **Descrição**: Otimizar dependências do useMemo para queries
+#### **1.3 Dependências Faltantes em useEffect** ⭐ PRIORIDADE ALTA
+- **Arquivo**: `src/components/RecurrenceConfig.jsx`
+- **Tempo estimado**: 2 horas
+- **Responsável**: Dev Pleno
+- **Descrição**: Corrigir dependências do useEffect para evitar loops infinitos
+- **Risco**: Médio - possíveis loops infinitos
+- **Mitigação**: Testar cada mudança isoladamente
 - **Implementação**:
   ```jsx
+  useEffect(() => {
+    // ... lógica
+  }, [calculateCurrentRule, onRuleChange]) // Adicionar onRuleChange
+  ```
+
+### **Sprint 2: Performance e Otimizações (10-12h)**
+
+#### **2.1 Re-renders Desnecessários em App.jsx** ⭐ PRIORIDADE ALTA
+- **Arquivo**: `src/App.jsx`
+- **Tempo estimado**: 8 horas (50% buffer incluído)
+- **Responsável**: Dev Senior
+- **Descrição**: Otimizar dependências do useMemo para queries, evitar recriação desnecessária
+- **Risco**: Alto - pode quebrar funcionalidade de queries
+- **Mitigação**: Testes incrementais, rollback plan
+- **Implementação**:
+  ```jsx
+  // ANTES: user como dependência completa
   const transactionsQuery = useMemo(
     () => user ? query(...) : null,
-    [db, appId, user?.uid] // Usar user.uid em vez de user
+    [db, appId, user] // user muda a cada render
+  )
+  
+  // DEPOIS: user.uid específico
+  const transactionsQuery = useMemo(
+    () => user?.uid ? query(...) : null,
+    [db, appId, user?.uid] // Apenas uid muda
   )
   ```
 
-#### **2.2 Função calculateCurrentRule Recriada**
+#### **2.2 Função calculateCurrentRule Recriada** ⭐ PRIORIDADE MÉDIA
 - **Arquivo**: `src/components/RecurrenceConfig.jsx`
 - **Tempo estimado**: 4 horas
 - **Responsável**: Dev Pleno
-- **Descrição**: Otimizar dependências do useCallback
+- **Descrição**: Otimizar dependências do useCallback para evitar recriação da função
+- **Risco**: Médio - possíveis re-renders em componentes filhos
+- **Mitigação**: Validar re-renders com React DevTools
 - **Implementação**:
   ```jsx
   const calculateCurrentRule = useCallback(() => {
@@ -92,17 +116,84 @@ Plano detalhado para implementação de **33 itens** (15 correções de bugs + 1
   }, [isEnabled, rule.type, rule.interval, rule.endDate, rule.count, endType])
   ```
 
-#### **2.3 Event Listeners Sem Cleanup Adequado**
+#### **2.3 Event Listeners Sem Cleanup Adequado** ⭐ PRIORIDADE MÉDIA
 - **Arquivo**: `src/views/DayDetailsView.jsx`
-- **Tempo estimado**: 3 horas
+- **Tempo estimado**: 4 horas
 - **Responsável**: Dev Pleno
-- **Descrição**: Estabilizar dependências com useCallback
+- **Descrição**: Estabilizar dependências com useCallback para evitar recriação de event listeners
+- **Risco**: Baixo - principalmente performance
+- **Mitigação**: Monitorar memory leaks
 - **Implementação**:
   ```jsx
   const handlePrevDay = useCallback(() => {
     // ... lógica
   }, [selectedDate, setSelectedDate])
   ```
+
+---
+
+## ⚠️ **RISCOS E MITIGAÇÕES - FASE 1**
+
+### **Risco Crítico: Firebase**
+- **Probabilidade**: Média
+- **Impacto**: Alto (quebra autenticação)
+- **Mitigação**: 
+  - Testar em branch separado
+  - Backup completo do código
+  - Rollback plan documentado
+  - Testes de autenticação completos
+
+### **Risco Alto: Performance**
+- **Probabilidade**: Alta
+- **Impacto**: Médio (bugs de performance)
+- **Mitigação**:
+  - Testes incrementais a cada mudança
+  - Medir performance antes/depois
+  - React DevTools para re-renders
+  - Console.log temporário para debug
+
+### **Risco Médio: Dependências**
+- **Probabilidade**: Média
+- **Impacto**: Médio (loops infinitos)
+- **Mitigação**:
+  - Testar cada mudança isoladamente
+  - Console.log temporário para validação
+  - React DevTools para useEffect
+  - Rollback imediato se detectar problema
+
+---
+
+## 📊 **CRONOGRAMA AJUSTADO - FASE 1**
+
+### **Semana 1: Sprint 1 - Estabilidade (8-10h)**
+- **Dia 1-2**: Firebase + Config (6h)
+- **Dia 3**: Logger + Dependências (4h)
+- **Dia 4-5**: Testes + Validação
+
+### **Semana 2: Sprint 2 - Performance (10-12h)**
+- **Dia 1-3**: App.jsx + Queries (8h)
+- **Dia 4-5**: useCallback + Event Listeners (4h)
+- **Dia 6**: Testes + Validação
+
+---
+
+## 🎯 **CRITÉRIOS DE SUCESSO - FASE 1**
+
+### **Antes da Fase 1**
+- Score qualidade: 7.2
+- Performance: Baseline
+- Bugs críticos: 6
+- Firebase: Múltiplas instâncias
+
+### **Após a Fase 1**
+- Score qualidade: 7.5+
+- Performance: 15-20% melhoria
+- Bugs críticos: 0
+- Firebase: Instância única
+- Logger: Padronizado
+- useEffect: Sem loops infinitos
+- useMemo: Otimizado
+- useCallback: Estável
 
 ---
