+++ src/contexts/AuthContext.jsx
@@ -1,22 +1,11 @@
 import React, { useState, useEffect, createContext, useContext } from 'react'
-import { initializeApp } from 'firebase/app'
 import {
-  getAuth,
   onAuthStateChanged,
   createUserWithEmailAndPassword,
   signInWithEmailAndPassword,
   signOut,
 } from 'firebase/auth'
-import { getFirestore } from 'firebase/firestore'
-
-const firebaseConfig = {
-  apiKey: import.meta.env.VITE_FIREBASE_API_KEY,
-  authDomain: import.meta.env.VITE_FIREBASE_AUTH_DOMAIN,
-  projectId: import.meta.env.VITE_FIREBASE_PROJECT_ID,
-  storageBucket: import.meta.env.VITE_FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: import.meta.env.VITE_FIREBASE_MESSAGING_SENDER_ID,
-  appId: import.meta.env.VITE_FIREBASE_APP_ID,
-}
+import { firebaseAuth, firebaseDb } from '../firebase/config'
 
 const AuthContext = createContext(null)
 
@@ -31,9 +20,9 @@ export default function AuthProvider({ children }) {
   const [user, setUser] = useState(null)
   const [loadingAuth, setLoadingAuth] = useState(true)
 
-  const app = initializeApp(firebaseConfig)
-  const auth = getAuth(app)
-  const db = getFirestore(app)
+  // Usar instâncias centralizadas do Firebase
+  const auth = firebaseAuth
+  const db = firebaseDb
 
   useEffect(() => {
     const unsubscribe = onAuthStateChanged(auth, (firebaseUser) => {
@@ -54,7 +43,7 @@ export default function AuthProvider({ children }) {
     loadingAuth,
     db,
     auth,
-    appId: firebaseConfig.appId,
+    appId: import.meta.env.VITE_FIREBASE_APP_ID,
     registerUser,
     loginUser,
     logoutUser,
